Move rendered content paste tests into paste.test.js
Consolidate paste-related tests into the existing paste test file
instead of creating a separate file.
